Make message constructor into a util
